Adicionar empresa e novos models

+++ APIHungryHunters/Models/Contas.cs
@@ -5,7 +5,7 @@ namespace APIHungryHunters.Models
     public class Contas
     {
         [Key]
-        public long Id_conta { get; set; }
+        public int Id_conta { get; set; }
 
         public string Email { get; set; }
         public string Username { get; set; }
